requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
requests==2.28.2
transformers==4.26.1
torch==1.13.1
pydantic==1.10.7
accelerate==0.18.0
python-dotenv